+++ packages/client/src/features/routing/index.ts
@@ -13,5 +13,5 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-export * from './glsp-manhattan-edge-router';
+export * from './edge-router';
 export * from './routing-module';
