+++ packages/glsp-sprotty/src/index.ts
@@ -16,6 +16,4 @@
 export * from './action-override';
 export * from './feature-modules';
 export * from './re-exports';
-export * from './routing-override';
 export * from './types';
-export * from './ui-extension-override';
