+++ packages/glsp-sprotty/src/re-exports.ts
@@ -74,7 +74,7 @@ export {
 } from 'sprotty/lib/base/model/smodel-factory';
 export * from 'sprotty/lib/base/model/smodel-utils';
 
-export { IUIExtension, isUIExtension } from 'sprotty/lib/base/ui-extensions/ui-extension';
+export * from 'sprotty/lib/base/ui-extensions/ui-extension';
 export * from 'sprotty/lib/base/ui-extensions/ui-extension-registry';
 
 export * from 'sprotty/lib/base/views/dom-helper';
@@ -131,7 +131,7 @@ export {
 } from 'sprotty/lib/features/command-palette/action-providers';
 export * from 'sprotty/lib/features/command-palette/command-palette';
 
-// Exclude menu item. Aready provided by glsp-protocol
+// Exclude menu item. Already provided by glsp-protocol
 export { Anchor, IContextMenuService, IContextMenuServiceProvider } from 'sprotty/lib/features/context-menu/context-menu-service';
 export { ContextMenuProviderRegistry, DeleteContextMenuItemProvider } from 'sprotty/lib/features/context-menu/menu-providers';
 export * from 'sprotty/lib/features/context-menu/mouse-listener';
@@ -194,25 +194,12 @@ export * from 'sprotty/lib/features/open/open';
 export * from 'sprotty/lib/features/projection/model';
 export * from 'sprotty/lib/features/projection/views';
 
-export {
-    DefaultAnchors,
-    LinearRouteOptions,
-    Side,
-    AbstractEdgeRouter as SprottyAbstractEdgeRouter
-} from 'sprotty/lib/features/routing/abstract-edge-router';
+export * from 'sprotty/lib/features/routing/abstract-edge-router';
 export * from 'sprotty/lib/features/routing/anchor';
 export * from 'sprotty/lib/features/routing/bezier-anchors';
-export {
-    AddRemoveBezierSegmentAction,
-    AddRemoveBezierSegmentCommand,
-    BezierMouseListener,
-    BezierEdgeRouter as SprottyBezierEdgeRouter
-} from 'sprotty/lib/features/routing/bezier-edge-router';
+export * from 'sprotty/lib/features/routing/bezier-edge-router';
 export * from 'sprotty/lib/features/routing/manhattan-anchors';
-export {
-    ManhattanRouterOptions,
-    ManhattanEdgeRouter as SprottyManhattanEdgeRouter
-} from 'sprotty/lib/features/routing/manhattan-edge-router';
+export * from 'sprotty/lib/features/routing/manhattan-edge-router';
 
 // Alias SModel types
 export {
@@ -230,7 +217,7 @@ export {
     isConnectable
 } from 'sprotty/lib/features/routing/model';
 export * from 'sprotty/lib/features/routing/polyline-anchors';
-export { PolylineRouteOptions, PolylineEdgeRouter as SprottyPolylineEdgeRouter } from 'sprotty/lib/features/routing/polyline-edge-router';
+export * from 'sprotty/lib/features/routing/polyline-edge-router';
 export * from 'sprotty/lib/features/routing/routing';
 export * from 'sprotty/lib/features/routing/views';
 
@@ -326,4 +313,3 @@ export * from 'sprotty/lib/utils/iterable';
 export * from 'sprotty/lib/utils/keyboard';
 export * from 'sprotty/lib/utils/logging';
 export * from 'sprotty/lib/utils/registry';
-
